Ok - updated the LWS function to include '\r' as well, because we were including it in our headers, and it was messing up the replies at certain point.
Also, and more important - I figured out that the reason that all the commands other than HEAD are allowed to wait for a body, and that is why we had the delays in responses.
Delete and PUT are allowed to have an empty body,
HEAD must NOT have a body,
and GET has to have the requested resource in the header.
I wrote a function to set a content-length header (currently its 0 - need to edit this) and now things work much much better.